Incipit
Started implementing a boundary matrix with the option of changing the
representation later on. This is inspired by PHAT [1] and their way of
using templates, which I very much admire.

While I do not plan to support that many algorithms, I think that this
sort of interface is very pretty.

What works so far:
  * Reading a boundary matrix, following the format of PHAT
  * Writing a boundary matrix to an std::ostream

Next: Providing a simple reduction algorithm. The code is already there
but I have not yet thought about how to represent the persistence pairs
in code.

[1]: https://bitbucket.org/phat-code/phat