Implemented basic calculation of persistence pairs
So far, I have not given any thought to the matter of storing the
persistence pairs. Ideally, this should be a rather simple data
structure. I am not sure at this point what operations are required,
though.

In any case, I now get the correct persistence pairs for a very simple
data set. This means that the standard reduction algorithm works as
expected now.