Extended test application with simplicial complex class
It is now possible to build a simplicial complex, bring it into some
form of filtration ordering, and calculate the persistent homology of
said filtration.

Nonetheless, several important aspects are still missing. I require a
better way to sort simplicial complexes. Furthermore, a representation
of the persistence pairs is needed. It should be convertible to a full
persistence diagram, or a set thereof.