Preliminary implementation for 1D function loading
This will be useful as a demonstration tool for certain properties of
persistent homology. In essence, I want to load vectors of function
values and convert them to a corresponding boundary matrix
automatically...

This requires some smart re-indexing if I do not want to use simplicial
complexes or other auxiliary data structures.