Phased out boundary matrix loading to boundary matrix class
This makes more sense because it is such an integral part of the
boundary matrix itself.

Although it could conceivably be implemented in terms of the public
interface of the class, boundary matrices are not supposed to be used
directly by most clients, making the need for more and more headers
obsolete.

This required some changes to the interface of other algorithms and
examples, though.

This fixes issue #4.