Added weighted random graph generation
In some sense, this is the weighted equivalent to the Erdős--Rényi
graphs. The difference is that this graph contains edge weights as
well. However, when using a 'forgetful functor' to make this graph
into an unweighted one, we obtain an Erdős--Rényi graph with equal
parameters $n$ and $p$.

Again, this is not yet tested.