Simplified calculation of topological centrality measure
This is slightly more efficient because I am not traversing *all* roots
any more. Next, I should try slicing the simplicial complex for an even
better performance---also, the sliced complex is guaranteed to have the
largest connected component I need to extract because I am accessing it
directly via the filtration instead of the values.