Removed automated face creation for simplicial complexes
In some cases, this may cause problems, but the present behaviour is
more true to the way the complex should be created. I could even add
a functor for the 'face creation phase' of the simplicial complex in
order to give clients more flexibility when choosing weights.

This solves issue #12.