Added support for "inline" level specification
I want to support GML files of the following form:

    graph [
      node [
        id A
      ]
    ]

Previously, the parse only understood the following format:

    graph
    [
      node
      [
        id A
      ]
    ]

In order to avoid having to convert those files, the parser is now
a little bit more standard-conforming. I also extended the tests a
little bit to demonstrate that the parser continues to work.