Started implementing data set item model
So far, this is more or less without any function. I am merely creating
a structure in which to store the data later on.

My concept behind this is to make everything data-centric: I want to
trigger interactions, such as the creation of a persistence diagram
based on views of the data.