Enumerating all perversities for the wicked triangulations
This is a precursor to the actual analysis I want to perform with Aleph,
namely the full enumeration of all Betti numbers for all possible values
of the perversity function.

So far, I am generating all combinations for the perversity correctly as
well as setting up skeletons for intersection homology.