Skeleton for basic matrix reader
I need a generic reader for matrix files in text format, i.e. files that
contain scalar data corresponding to row/column entries in a matrix. The
classical example of one of these files is a greyscale image. While it's
certainly possible to implement readers for *generic* image formats such
as PNG, this way is simpler and more flexible as we rely on outside data
conversion tools.