Simplified test variable configuration
It makes no sense to keep these in a *global* configuration file that
might be part of the installation. Instead, these variables should be
kept in a configuration file that is only accessible from the testing
suite.

Added a .gitignore file so that `Base.hh` is not added by accident.