Fixed admissibility calculation for intersection homology
Using problematic placement of brackets, the admissibility calculation
was not working as expected. I caught this because of the *unit tests*
and first thought it was related to the early abort procedure. Instead
it was a combination of an issue with the boundary matrix (fixed), and
the way the boolean expression was being evaluated.