Preparing test case for intersection homology dualization
This turns out to be more complicate than expected: simplices with
higher indices must *not* be used here. This makes it harder for a
boundary matrix to be dualized, because every reduction algorithm,
at least the way they are implemented right now, tacitly needs the
matrix to be *square*. Or else, they are reducing the wrong column
during the operation. Probably need to fully rethink this.