Providing additional standard reduction algorithm for rectangular matrices
I am little bit torn concerning a proper solution for this issue. On the
one hand, I do not want to waste cycles by adding an additional check as
for quadratic matrices, this is *not* necessary. On the other hand, code
duplication is not nice as well.

For now, leaving this as-is and re-implementing the reduction schemes in
order to finally get dual matrices to work.