Keeping track of destroyers in persistence pairing calculation
The reason for this is that dualized boundary matrices that do not
coincide with the filtration order may cause problems with respect
to the order in which the pairing is being calculated. Hence, I am
now tracking destroyers to ensure that they are not used twice. As
unpaired creators may cause troubles later on, this is useful when
calculating Betti numbers, for example.