Added test data set for $S^1 \vee S^1$
Need this in order to debug/visualize/understand different variants of
the spine creation algorithm. This data set has been created by adding
eight points on a unit circle, duplicating the circle, and adding four
points around the centre of the object; they are supposed to represent
noise in the data.