Added sanity check for Python bindings
This fixes issue #43 and ensures that users are at least *warned* if
their Python interpreter and their include directories do not match.

Notice that this will not detect all inconsistencies, alas. There is
still the problem of taking a Python 3 interpreter with strange sets
if libraries, for example an Anaconda installation using the default
system libraries.